// src/app/shared/models/faqs.ts
export class Faqs {
    categoryId: number
    type: number
    id: number
    answer: string
    question: string
    categoryName: string
    displayOrder: number
    faqTypeId: number
    faqType: number
    faqCategoryId: number
    isActive: boolean
    isDelete: boolean
    createdAt: Date
    updatedAt: Date
    createdBy: number
    modifiedBy: number
    name:string
    constructor() { }
}